+++ llm_quest/vit/vit_engine.py
@@ -88,7 +88,7 @@ def vit_training_eval_loop(
             targets = targets.to(device)
 
             # Autocast enable/disable for mixed precision training
-            with torch.autocast("cuda", enabled=use_amp):
+            with torch.autocast("cuda", dtype=torch.bfloat16, enabled=use_amp):
                 logits = model(input_batch)
                 loss = torch.nn.functional.cross_entropy(logits, targets)
